Use original declaring class for Method signature
As `Method#declaringClass` refers to the type that actually defines the
method and not the class that was scanned for method.

In the case of public methods inherited from super class or interfaces,
the compatibility checker needs to know the original class that was
scanned even if some methods are declared in super types.

This is for filtering signatures and displaying them in errors.